requirements.txt
streamlit
huggingface_hub
Pillow
python-dotenv